% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.r
\name{create_folder}
\alias{create_folder}
\title{Create folders}
\usage{
create_folder(path, token = NULL)
}
\arguments{
\item{path}{remote path starting in your home directory.}

\item{token}{token. Use \code{login()} to connect to a server or provide token
directly here.}
}
\description{
Create folders
}
\examples{
\dontrun{
list_files()
}
}
