% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{delete_files}
\alias{delete_files}
\title{Delete remote files}
\usage{
delete_files(path, token = NULL, verbose = TRUE)
}
\arguments{
\item{path}{remote path starting in your home directory.}

\item{token}{token. Use \code{login()} to connect to a server or provide token
directly here.}

\item{verbose}{print status to screen?}
}
\description{
Delete remote files
}
\examples{
\dontrun{
list_files("Photos")
delete_files("/Photos/Vineyard.jpg")
}
}
