% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{download_files}
\alias{download_files}
\title{Download remote files}
\usage{
download_files(
  path = "",
  destination = basename(path),
  token = NULL,
  verbose = TRUE
)
}
\arguments{
\item{path}{remote path starting in your home directory.}

\item{destination}{local file path to download to.}

\item{token}{token. Use \code{login()} to connect to a server or provide token
directly here.}

\item{verbose}{print status to screen?}
}
\description{
Download remote files
}
\examples{
\dontrun{
list_files("Photos")
download_files("/Photos/Vineyard.jpg")
}
}
