% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{list_files}
\alias{list_files}
\title{List remote files}
\usage{
list_files(path = "", token = NULL, verbose = TRUE)
}
\arguments{
\item{path}{remote path starting in your home directory.}

\item{token}{token. Use \code{login()} to connect to a server or provide token
directly here.}

\item{verbose}{print status to screen?}
}
\value{
a data.frame of files and metadata
}
\description{
List remote files
}
\examples{
\dontrun{
list_files()

# list files in folder
list_files("Photos")
# OR
list_files("/Photos/")
}
}
