% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{login}
\alias{login}
\title{Login to a server}
\usage{
login(server, user = NULL, use_cached = TRUE)
}
\arguments{
\item{server}{the URL to a 'Nextcloud' server.}

\item{user}{optional. The user name (only relevant if multiple user tokens
are cached).}

\item{use_cached}{\code{FALSE} forces the creation of a new token.}
}
\value{
token (invisible)
}
\description{
Login to a server
}
\examples{
\dontrun{
login("https://cloud.example.com")
}
}
