% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextcloudr-package.R
\docType{package}
\name{nextcloudr-package}
\alias{nextcloudr}
\alias{nextcloudr-package}
\title{nextcloudr: Connect to a 'Nextcloud' server}
\description{
Implementation of the WebDAV operations supported in 'Nextcloud'.
}
\author{
\strong{Maintainer}: Johannes B. Gruber \email{JohannesB.Gruber@gmail.com} (\href{https://orcid.org/0000-0001-9177-1772}{ORCID})

}
\keyword{internal}
