% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.r
\name{create_folder}
\alias{create_folder}
\title{Create folders}
\usage{
create_folder(path, recursive = FALSE, error_on_exist = FALSE, token = NULL)
}
\arguments{
\item{path}{remote path starting in your home directory.}

\item{recursive}{create folders recursively. Default is \code{FALSE} for
performance.}

\item{error_on_exist}{throw an error when the folder already exists.}

\item{token}{token. Use \code{login()} to connect to a server or provide token
directly here.}
}
\description{
Create folders
}
\examples{
\dontrun{
list_files()
}
}
