% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_files}
\alias{upload_files}
\title{Upload local files}
\usage{
upload_files(source, path, overwrite = TRUE, token = NULL, verbose = TRUE)
}
\arguments{
\item{source}{local file to upload.}

\item{path}{remote path starting in your home directory.}

\item{overwrite}{do you want to overwrite existing files (\code{TRUE}) or skip
them (\code{FALSE})?}

\item{token}{token. Use \code{login()} to connect to a server or provide token
directly here.}

\item{verbose}{print status to screen?}
}
\description{
Upload local files
}
\examples{
\dontrun{
jpgs <- list.files(pattern = "jpg")
upload_files(
  source = jpgs,
  path = file.path("/InstantUpload/", jpgs),
  overwrite = FALSE
)
}
}
