% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_files}
\alias{upload_files}
\title{Upload local files}
\usage{
upload_files(source, path, token = NULL)
}
\arguments{
\item{source}{local file to upload.}

\item{path}{remote path starting in your home directory.}

\item{token}{token. Use \code{login()} to connect to a server or provide token
directly here.}
}
\description{
Upload local files
}
\examples{
\dontrun{
list_files()
}
}
